0.0.6 big remake! & command line tools are more capable!
a new cmd_parser will now take over the job to analyse, breakdown, check correctness of the invoking command of m2l

introduced a new configurehandler.py and a new helpme, but these 2 will be implemented later

inputhandler is now iohandler

a new workflow will now provide a process build the workflow from given provider and consumers. The idea was: "provider" provides a string of the markdown file; "process" takes that in, and return the target tex string; "consumers" which may be multiple, outputs the tex string in different ways.